+++ code/src/sixsq/nuvla/server/resources/callback_create_user_mitreid.clj
@@ -33,7 +33,7 @@
                     (ex/create-user! :mitreid {:instance       instance
                                                :external-id    sub
                                                :external-email email})
-                    (oidc-utils/throw-user-exists sub redirect-url))
+                    (oidc-utils/throw-user-exists redirect-url))
                   (oidc-utils/throw-no-email redirect-url)))
               (oidc-utils/throw-no-subject redirect-url)))
           (catch Exception e