+++ code/src/sixsq/nuvla/server/resources/callback_create_user_oidc.clj
@@ -30,7 +30,7 @@
                 (ex/create-user! :oidc {:instance       instance
                                         :external-id    sub
                                         :external-email (or email (str sub "@fake-email.com"))})
-                (oidc-utils/throw-user-exists sub redirect-url))
+                (oidc-utils/throw-user-exists redirect-url))
               (oidc-utils/throw-no-subject redirect-url)))
           (catch Exception e
             (oidc-utils/throw-invalid-access-code (str e) redirect-url)))