+++ code/src/sixsq/nuvla/server/resources/github/utils.clj
@@ -21,8 +21,8 @@
 (defn throw-no-matched-user [redirect-url]
   (logu/log-error-and-throw-with-redirect 403 "no matching account for GitHub user" redirect-url))
 
-(defn throw-user-exists [username redirect-url]
-  (logu/log-error-and-throw-with-redirect 400 (str "account already exists (" username ")") redirect-url))
+(defn throw-user-exists [redirect-url]
+  (logu/log-error-and-throw-with-redirect 400 "an account with the same email already exists. Have you already registered with email/password?" redirect-url))
 
 (defn config-github-params
   [redirect-url instance]