+++ code/test/sixsq/nuvla/server/resources/deployment_lifecycle_test.clj
@@ -154,9 +154,13 @@
                                     (ltu/is-operation-present :edit)
                                     (ltu/is-operation-present :delete)
                                     (ltu/is-operation-present :start)
+                                    (ltu/is-operation-present :clone)
+                                    (ltu/is-operation-present :fetch-module)
                                     (ltu/is-key-value :state "CREATED"))
 
             start-url           (ltu/get-op-url deployment-response "start")
+
+            fetch-module-url    (ltu/get-op-url deployment-response "fetch-module")
 
             deployment          (ltu/body deployment-response)]
 
@@ -176,7 +180,31 @@
             (is (:description credential))
             (is (= deployment-id (:parent credential)))
 
+            (let [module (-> session-user
+                             (request (str "/api/" module-id))
+                             (ltu/body->edn)
+                             (ltu/is-status 200)
+                             :response :body)]
 
+              (-> session-user
+                 (request (str "/api/" module-id)
+                          :request-method :put
+                          :body (json/write-str (assoc-in module [:content :environmental-variables]
+                                                          [{:name "ALPHA_ENV", :value "NOK"}
+                                                           {:name "NEW", :value "new"}
+                                                           {:name "BETA_ENV",
+                                                            :description "beta-env variable",
+                                                            :required true}]
+                                                          )))
+                 (ltu/body->edn)
+                 (ltu/is-status 200)))
+
+            ;; try call fetch-module
+            (-> session-user
+                (request fetch-module-url
+                         :request-method :post)
+                (ltu/is-status 200)
+                (ltu/body->edn))
 
             ;; attempt to start the deployment and check the start job was created
             (let [job-url (-> session-user
@@ -319,7 +347,9 @@
                         (request deployment-url)
                         (ltu/body->edn)
                         (ltu/is-status 200)
-                        (ltu/is-key-value :state "STOPPING"))
+                        (ltu/is-key-value :state "STOPPING")
+                        (ltu/is-operation-absent "fetch-module")
+                        (ltu/is-operation-absent "start"))
 
                     ;; the deployment would be set to "STOPPED" via the job
                     ;; for the tests, set this manually to continue with the workflow
@@ -329,6 +359,41 @@
                                  :body (json/write-str {:state "STOPPED"}))
                         (ltu/body->edn)
                         (ltu/is-status 200))
+
+                    ;; stopped deployment can be started again
+                    (-> session-user
+                        (request deployment-url)
+                        (ltu/body->edn)
+                        (ltu/is-status 200)
+                        (ltu/is-operation-present "start"))
+
+                    ;; verify user can create another deployment from existing one
+                    (let [deployment-url-from-dep (-> session-user
+                                                      (request base-uri
+                                                               :request-method :post
+                                                               :body (json/write-str {:deployment {:href deployment-id}}))
+                                                      (ltu/body->edn)
+                                                      (ltu/is-status 201)
+                                                      (ltu/location-url))]
+                      (-> session-user
+                          (request deployment-url-from-dep
+                                   :request-method :delete)
+                          (ltu/body->edn)
+                          (ltu/is-status 200)))
+
+                    ;; verify user can create another deployment from existing one by using clone action
+                    (let [deployment-url-from-dep (-> session-user
+                                                      (request (str deployment-url "/clone")
+                                                               :request-method :post
+                                                               :body (json/write-str {:deployment {:href deployment-id}}))
+                                                      (ltu/body->edn)
+                                                      (ltu/is-status 201)
+                                                      (ltu/location-url))]
+                      (-> session-user
+                          (request deployment-url-from-dep
+                                   :request-method :delete)
+                          (ltu/body->edn)
+                          (ltu/is-status 200)))
 
                     ;; verify that the user can delete the deployment
                     (-> session-user
@@ -367,7 +432,7 @@
                      :body (json/write-str invalid-deployment))
             (ltu/body->edn)
             (ltu/is-status 400)
-            (ltu/message-matches #"cannot resolve module .*"))
+            (ltu/message-matches #"cannot resolve.*"))
         ))
 
     (-> session-user