+++ code/test/sixsq/nuvla/server/resources/user_template_mitreid_lifecycle_test.clj
@@ -341,7 +341,7 @@
                       (request (str val-url "?code=GOOD")
                                :request-method :get)
                       (ltu/body->edn)
-                      (ltu/message-matches #".*account already exists.*")
+                      (ltu/message-matches #".*an account with the same email already exists.*")
                       (ltu/is-status return-code))
 
                   (is (= "FAILED" (-> session-admin